Rerender navbar on route changes
This commit adds changes in order to update the Navbar when
the user's login state changes.

I believe a better approach would be to add an observer for changes
in the Auth state and pass that as a singleton to components that
should update when the Auth state changes. However, implementing
this change would require changes in several components.

This change updates only the Navbar and /client/src/index.js by
modifying the Navbar so that it listens for changes in the history
prop injected by BrowserRouter and updating the Navbar component
if the Auth state has changed.