Refactoring client to extract Home component
This commit renames the App component to pages/Home/index.js so that the
name better represents the purpose of the component and to better align
the components with activities students have experienced in class.

The routing and other global components have been moved from
src/index.js to App.js.

The sole concern of /src/index.js is startup initialization,
configuration and mounting the App to the DOM.